backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
numpy
apscheduler
requests
python-dotenv
psycopg2-binary